# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/messages/contacts": {
      "get": {
        "operationId": "MessagingController_contacts",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/v1/messages/conversations": {
      "get": {
        "operationId": "MessagingController_list",
        "parameters": [
          {
            "name": "patientId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "MessagingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}": {
      "get": {
        "operationId": "MessagingController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}/messages": {
      "get": {
        "operationId": "MessagingController_messages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "description": "Older than this time (the `createdAt` of the oldest message you have).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "MessagingController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}/read": {
      "post": {
        "operationId": "MessagingController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}/participants": {
      "post": {
        "operationId": "MessagingController_addParticipants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}/leave": {
      "post": {
        "operationId": "MessagingController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Alora Health API",
    "description": "Home health agency management API. All routes are under `/api/v1`.\n\nSuccess responses are wrapped as `{ \"success\": true, \"data\": …, \"meta\"?: { page, limit, total } }`.\nErrors are `{ \"success\": false, \"error\": { \"code\", \"message\", \"details\"? } }`.\n\nAuthenticate with `POST /api/v1/auth/login`, then send `Authorization: Bearer <accessToken>`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateConversationDto": {
        "type": "object",
        "properties": {
          "participantIds": {
            "description": "The other people (the caller is added automatically).",
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "subject": {
            "type": "string",
            "maxLength": 255
          },
          "patientId": {
            "type": "string",
            "description": "The patient this is about (the caller must have access to them).",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "maxLength": 5000
          },
          "isUrgent": {
            "type": "boolean"
          },
          "documentId": {
            "type": "string",
            "description": "A document the sender can see (uploaded through /documents).",
            "format": "uuid"
          }
        },
        "required": [
          "participantIds",
          "content"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 5000
          },
          "isUrgent": {
            "type": "boolean"
          },
          "documentId": {
            "type": "string",
            "description": "A document the sender can see (uploaded through /documents).",
            "format": "uuid"
          }
        },
        "required": [
          "content"
        ]
      },
      "AddParticipantsDto": {
        "type": "object",
        "properties": {
          "userIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "userIds"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}